Car Price Prediction-Regression Analysis
With the rise of the vehicle business, car price prediction has become a popular study topic. The number of automobiles on the road has been rising, and this has increased the popularity of this subject. A huge variety of diverse attributes play a vital part in a dependable and accurate forecast of the price of a car. Fuel type, engine type, car
dimensions, among other factors, will be investigated in this project in order to gain a better knowledge of this subject and identify the key features that majorly affect the priceof a car. The key features identified will be used to build a model that will be help to predict
the car price interval when the key features are provided